SlotNumber/MagnetStatic int->byte
more size reductions

+++ PKHeX.Core/Legality/Areas/EncounterArea3.cs
@@ -68,14 +68,14 @@ namespace PKHeX.Core
         {
             int species = ReadUInt16LittleEndian(entry);
             int form = entry[2];
-            int slotNum = entry[3];
+            byte slotNum = entry[3];
             int min = entry[4];
             int max = entry[5];
 
-            int mpi = entry[6];
-            int mpc = entry[7];
-            int sti = entry[8];
-            int stc = entry[9];
+            byte mpi = entry[6];
+            byte mpc = entry[7];
+            byte sti = entry[8];
+            byte stc = entry[9];
             return new EncounterSlot3(this, species, form, min, max, slotNum, mpi, mpc, sti, stc);
         }
 
@@ -98,7 +98,7 @@ namespace PKHeX.Core
         {
             int species = ReadUInt16LittleEndian(entry);
             // form always 0
-            int slotNum = entry[3];
+            byte slotNum = entry[3];
             int min = entry[4];
             int max = entry[5];
