+++ PKHeX.Core/Legality/Encounters/EncounterSlot/EncounterSlot1.cs
@@ -7,10 +7,10 @@
     public sealed record EncounterSlot1 : EncounterSlot, INumberedSlot
     {
         public override int Generation => 1;
-        public int SlotNumber { get; }
+        public byte SlotNumber { get; }
         public override Ball FixedBall => Ball.Poke;
 
-        public EncounterSlot1(EncounterArea1 area, int species, int min, int max, int slot) : base(area, species, 0, min, max)
+        public EncounterSlot1(EncounterArea1 area, int species, int min, int max, byte slot) : base(area, species, 0, min, max)
         {
             SlotNumber = slot;
         }
