+++ PKHeX.Core/Legality/Encounters/EncounterSlot/EncounterSlot3.cs
@@ -8,16 +8,16 @@ namespace PKHeX.Core
     {
         public sealed override int Generation => 3;
 
-        public int StaticIndex { get; }
-        public int MagnetPullIndex { get; }
-        public int StaticCount { get; }
-        public int MagnetPullCount { get; }
+        public byte StaticIndex { get; }
+        public byte MagnetPullIndex { get; }
+        public byte StaticCount { get; }
+        public byte MagnetPullCount { get; }
         public SlotType Type => Area.Type;
 
-        public int SlotNumber { get; }
+        public byte SlotNumber { get; }
         public override Ball FixedBall => Locations.IsSafariZoneLocation3(Location) ? Ball.Safari : Ball.None;
 
-        public EncounterSlot3(EncounterArea3 area, int species, int form, int min, int max, int slot, int mpi, int mpc, int sti, int stc) : base(area, species, form, min, max)
+        public EncounterSlot3(EncounterArea3 area, int species, int form, int min, int max, byte slot, byte mpi, byte mpc, byte sti, byte stc) : base(area, species, form, min, max)
         {
             SlotNumber = slot;
