+++ PKHeX.Core/Legality/Encounters/EncounterSlot/EncounterSlot3Swarm.cs
@@ -13,7 +13,7 @@ namespace PKHeX.Core
     {
         public IReadOnlyList<int> Moves { get; }
 
-        public EncounterSlot3Swarm(EncounterArea3 area, int species, int min, int max, int slot,
+        public EncounterSlot3Swarm(EncounterArea3 area, int species, int min, int max, byte slot,
             IReadOnlyList<int> moves) : base(area, species, 0, min, max, slot, 0, 0, 0, 0) => Moves = moves;
 
         protected override void SetEncounterMoves(PKM pk, GameVersion version, int level)
